Add files via upload
Image sets of calibrators and lateral flow tests. Lateral flow file names are arranged as either pre-covid healthy controls or vaccinated healthy controls. The second part of the file names describes the dilution factor with PBS, for example 1to8. Each test was photographed 3 times, and the final part of the file name refers to which number the photo is in the subset. 
Example: vaccinated1to2.3.jpg describes that the plasma was from a healthy vaccinated control, the dilution of plasma to PBS was 1:2, and this is the third image in the set.